% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfunction.R
\name{vfunction}
\alias{vfunction}
\title{create an R wrapper from a v function.}
\usage{
vfunction(sig, body)
}
\arguments{
\item{sig}{\code{character}}

\item{body}{\code{character}}
}
\description{
like \code{inline::cfunction}, but for \code{v}
}
