% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_export_c.R
\name{rv_export_c}
\alias{rv_export_c}
\title{Generate rv_export files for .c}
\usage{
rv_export_c(pkg = ".")
}
\arguments{
\item{pkg}{\code{character} location of package.}
}
\description{
Generates rv_export files from a \code{v} source file, that is located in the
\code{src} directory of a pkg.
}
\details{
The following files are generated:
\itemize{
\item "./R/rv_export.R": R functions that wrap the v functions
\item "./src/rv_export.v": v wrappers, that interface v with R.
\item "./src/init.c" : registration code, needed for the shared library.
\item "./src/<pkg_name>.c" : registration code, needed for the shared library.
}
}
